src/requirements.txt
torch>=1.12.0
torchvision
timm>=0.4.12
tqdm
scikit-learn
pandas
numpy
Pillow

